# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Lunme API",
    "version": "1.0.0",
    "description": "Swagger gerado automaticamente pelas rotas do Next.js"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/internal/gerar-link-anamnese": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Gera um link temporário de acesso à anamnese",
        "description": "Cria um token temporário para o psicólogo acessar a anamnese de forma segura e retorna o link completo.",
        "tags": [
          "Interno - Anamnese"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "psicologoId"
                ],
                "properties": {
                  "psicologoId": {
                    "type": "string",
                    "example": "psico_98765"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Link gerado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "link": {
                      "type": "string",
                      "example": "https://meusite.com/amnp/psico_98765/123e4567-e89b-12d3-a456-426614174000"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros obrigatórios não informados.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Campos obrigatórios: psicologoId"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor ao gerar o link.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro interno do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/history": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Adiciona um registro ao histórico de um psicólogo",
        "description": "Cria um novo registro no histórico de um psicólogo, com descrição, tipo e timestamp.",
        "tags": [
          "Interno - Histórico"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "psicologoId",
                  "descricao"
                ],
                "properties": {
                  "psicologoId": {
                    "type": "string",
                    "example": "psico_98765"
                  },
                  "descricao": {
                    "type": "string",
                    "example": "Consulta realizada com sucesso"
                  },
                  "tipo": {
                    "type": "string",
                    "example": "geral"
                  },
                  "timestamp": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-11-20T14:30:00Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Histórico criado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "hist_123456"
                    },
                    "psicologoId": {
                      "type": "string"
                    },
                    "descricao": {
                      "type": "string"
                    },
                    "tipo": {
                      "type": "string"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros obrigatórios não informados.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "psicologoId e descricao são obrigatórios"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno ao salvar histórico.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro interno ao salvar histórico"
                    },
                    "details": {
                      "type": "string",
                      "example": "Mensagem de erro detalhada do servidor"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Lista o histórico de um psicólogo",
        "description": "Retorna todos os registros do histórico de um psicólogo, ordenados do mais recente para o mais antigo.",
        "tags": [
          "Interno - Histórico"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "psicologoId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do psicólogo cujos registros serão retornados."
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico retornado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "hist_123456"
                      },
                      "psicologoId": {
                        "type": "string"
                      },
                      "descricao": {
                        "type": "string",
                        "example": "Consulta realizada com sucesso"
                      },
                      "tipo": {
                        "type": "string",
                        "example": "geral"
                      },
                      "timestamp": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-20T14:30:00Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Parâmetro psicologoId não informado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "psicologoId é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao buscar histórico.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao buscar histórico"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Deleta um registro do histórico de um psicólogo",
        "description": "Remove um registro do histórico com base no ID fornecido.",
        "tags": [
          "Interno - Histórico"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID do registro de histórico a ser deletado."
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico deletado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Histórico deletado com sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID do histórico não informado.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "ID do histórico é obrigatório"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro ao deletar o histórico.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Erro ao deletar histórico"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": []
}